Support "--dev" flag for development mode
I got the external volume mount sort of wrong. We only want to mount the
source and sound effects, not node_modules or configuration. I've split
this off into a separate mode that one only runs if specifically
desired.

Also added some small adjustments to the bot itself. Namely moved it to
The Reader channel and moved channel names to constants. Will need to
make this configurable instead of hard-coded.